Update gollum repository location in the Gemfile.
Github just moved the Gollum repository to it's own organization.
This commit updates the Gemfile to point to the new repo location.

I also added a specific ref pointer to a version that is known to
work well just in case they slip up and break their master branch
right at the same time someone happens to be installing or updating
Gitlab. This is about the safest we can get until they put out the
next release that resolves the Pygments dependency issue.