Wording improvements in the zero_authorized_projects template
I've reworded the first sentence to say "You don't have access to any projects",
which I think sounds more natural.

I've also used the "pluralize" ActionView helper for the user's projects_limit,
so that it still looks good when it says "You can create up to 1 project.".

Finally, I've tweaked the structure of the final sentence a little bit to make
it sound a bit better.